Move RegistrationOptions.cs to Config folder & switch  from IConfigure to IOptions
